+++ pyflp/exceptions.py
@@ -39,6 +39,8 @@ class Error(Exception):
 
 
 class EventIDOutOfRange(Error, ValueError):
+    """An event is created with an ID out of its allowed range."""
+
     def __init__(self, id: int, min_i: int, max_e: int):
         super().__init__(f"Expected ID in {min_i}-{max_e - 1}; got {id} instead")
 
@@ -48,6 +50,10 @@ class InvalidEventChunkSize(Error, TypeError):
 
     def __init__(self, expected: int, got: int):
         super().__init__(f"Expected a bytes object of length {expected}; got {got}")
+
+
+class ListEventNotParsed(Error, AttributeError, IndexError):
+    """`ListEventBase` could not be parsed due to mismatching struct layouts."""
 
 
 class UnexpectedType(Error, TypeError):